fix caching issue (needs single file due to a nextjs issue being unable to maintain one instance of a module)
