+++ morpheus-DFP/package.json
@@ -20,6 +20,7 @@
     "echarts-for-react": "^3.0.2",
     "next": "^13.0.0",
     "next-transpile-modules": "^9.1.0",
+    "node-cache": "5.1.2",
     "rc-slider": "^10.0.1",
     "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
